refactor: difficulty menu
-Refactor code to reduce redundancy.
